moving option to launch server config modal into the admin tab of settings
