+++ js/templates/settings.html
@@ -1129,7 +1129,25 @@
                 </div>
               </div>
 
-            </div>
+            </div> 
+
+            <div class="flexRow">
+
+              <div class="flexCol-3  borderRight custCol-border">
+                <div class="fieldItem">
+                  <label>
+                    Server Settings
+                  </label>
+                </div>
+              </div>
+
+              <div class="flexCol-9 borderRight0 custCol-border">
+                <div class="fieldItem">
+                  <a class="btn btn-txt js-changeServerSettings">Change</a>
+                </div>
+              </div>
+
+            </div>                       
 
             <a class="btn btn-large js-saveAdvanced custCol-secondary pull-right marginRight15 marginTop15 marginBottom50 btn-secondary">
               <span class="ion-checkmark fontSize11 marginRight2 textOpacity1"></span>
